\name{graphVis3DSphere}
\alias{graphVis3DSphere}
\title{Graph Vis 3D Sphere}
\usage{
graphVis3DSphere(graph, roiNames, cutoff = NULL, expansion = 1.2,
  bend = 0.5, lwd = 2, col = "black")
}
\arguments{
  \item{graph}{graph as a matrix or vector}

  \item{roiNames}{a character vector of the roi names}

  \item{expansion}{how far points are from the center}

  \item{bend}{how bendy the lines are}

  \item{lwd}{line width}

  \item{col}{color}
}
\description{
a function to display a graph on a set of 3D spheres
}
\author{
Brian Caffo
}

