\name{templateBrainGraphVis}
\alias{templateBrainGraphVis}
\title{Template Brain Graph visualization}
\usage{
templateBrainGraphVis(pvalues, cutoff, templateImage, roiImage, roiNumbers,
  roiNames, lwd = 2, col = "lightblue")
}
\description{
function to plot a graph on a template brain roiNumbers
should be in unique(roiImage) are the integer image values
to draw the graph at. roiNumbers should be the same length
as pvalues.
}
\author{
Brian Caffo
}

